rename(feature/#46): api 관련 함수들 apis 폴더로 이동
