+++ src/apis/user.ts
@@ -1,6 +1,17 @@
-import { AuthUserInfo, UpdateUserFn, User } from '@/types/User'
-import { createSupabaseBrowserClient } from '../client/supabase'
-import { getRandomNickname } from './nicknameList'
+import { AuthUserInfo, UpdateUserFn } from '@/types/User'
+import { createSupabaseBrowserClient } from '@/utils/client/supabase'
+import { getRandomNickname } from '@/utils/user/nicknameList'
+
+/**
+ * 현재 로그인한 사용자의 id를 가져오는 함수
+ * @returns {Promise<string | void>} - 로그인 상태인 경우 유저 id, 실패 시 콘솔에 오류 메시지 출력
+ */
+export const getUserId = async () => {
+  const supabase = createSupabaseBrowserClient()
+  const { data, error } = await supabase.auth.getUser()
+  if (error) return console.error('Failed to find user:', error.message)
+  return data.user.id
+}
 
 /**
  * [GET] 사용자 정보를 가져오는 함수
@@ -15,6 +26,7 @@ export const getUserInfo = async (id: string) => {
   if (error) return console.error('failed to fetch user info', error.message)
   return data
 }
+
 /**
  * 특정 유저의 is_all_clear 값을 가져오는 함수.
  * @param {string} id - 조회할 유저의 ID.
@@ -64,7 +76,6 @@ export const createUser = async (authUserInfo: AuthUserInfo) => {
  * @param {string | number} changeContent - 수정할 값
  * @returns {Promise<boolean>} - 수정 성공 여부
  */
-
 export const updateUser: UpdateUserFn = async (user_id, target, changeContent) => {
   const supabase = await createSupabaseBrowserClient()
