+++ src/app/mypage/page.tsx
@@ -1,11 +1,10 @@
 'use client'
 import CategoryField from '@/components/common/CategoryField'
 import UserProfile from '@/components/user/UserProfile'
-import { getUserId } from '@/utils/user/auth'
-import { getUserInfo } from '@/utils/user/user'
 import { useEffect, useState } from 'react'
 import { getCompletedCategoryList } from '@/apis/category'
 import { Category, CategoryName } from '@/types/CategoryField'
+import { getUserId, getUserInfo } from '@/apis/user'
 
 export default function MyPage() {
   const [user, setUser] = useState({ user_id: '', nickname: '', avatar_url: '' })
@@ -38,7 +37,7 @@ export default function MyPage() {
 
   return (
     <main className="w-full h-[calc(100%-44px)] flex items-center flex-col p-[35px] overflow-hidden">
-      <div className="w-full flex items-center gap-[15px]">
+      <div className="w-full flex items-center gap-[15px] mt-[44px] mx-[35px]">
         <UserProfile size={50} isEditable={false} imgSrc={user.avatar_url} />
         <span className="font-gothic-m text-brown text-[12px] whitespace-nowrap">
           {user.nickname || '구해줘요 동물의 숲'}님<br /> 현재 뱃지 {categoryList.length}개를
